第一阶段结束

+++ NavMeshGenVisual/Assets/Script/NMGen/SolidHeightfieldBuilder.cs
@@ -27,7 +27,7 @@ namespace NMGen
             float cellHeight,
             int minTraversableHeight ,
             int maxTraversableStep,
-            int maxTraversableSlope,
+            float maxTraversableSlope,
             bool clipLedges
             )
         {
